% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LITree.R
\name{em.litree}
\alias{em.litree}
\title{em.litree}
\usage{
em.litree(X, k = 0:2, criterion = "ICL_T", max.iter = 20, eps = 0.1)
}
\arguments{
\item{X}{X is a data matrix}

\item{k}{is a vector of missing variable number. It should be an integer greater than 0 or a vector of such integers}

\item{criterion}{is the name of the criterion used for selecting the best model ({"ICL_T", "ICL_ZT", "BIC" }). ICL_T by default}

\item{max.iter}{is the maximum number of iterations (20 by default)}

\item{eps}{is the precision used for stopping the algorithm}
}
\value{
A list of three items, \code{criteria} a dataframe whose columns are the three critera and each lines corresponds
to a given number of missing variables, a list of models and the best model according the specified criterion
}
\description{
The function \code{em.litree} implements the algorithm of Gaussian Graphical Model Inference with missing variable described
in Robin et. al (2017). The underlying model is based on the aggregation of spanning trees,
 and the estimation procedure on the Expectation-Maximization algorithm.
 We treat the graph structure and the unobserved nodes as missing variables and compute posterior probabilities of edge appearance.
To provide a complete methodology, we also propose three model selection criteria to estimate the number of missing nodes.
}
\examples{
data(cyto)
res.raf.full <- em.litree(X.raf,0:2)

}
\references{
Geneviève Robin, Christophe Ambroise, Stéphane Robin (Submitted on 26 May 2017).
Graphical model inference with unobserved variable via latent tree aggregation. Arxiv Paper.
\url{https://arxiv.org/abs/1705.09464}
}
