% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LITree-class.R
\docType{data}
\name{graphModel}
\alias{graphModel}
\alias{graph.model}
\title{R6 Class for simulation of Graphs}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
graph.model
}
\description{
R6 Class for simulation of Graphs
}
\keyword{datasets}
