% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LITree-class.R
\name{GGMexperiment}
\alias{GGMexperiment}
\title{R6 Class for running experiment of Gaussian Graphical Model}
\description{
A simple example of what is possible using roxygen 4.1.1
}
\section{Usage}{

\preformatted{experiment = GGM.experiment$new(methods=c("glasso","em.latent.trees"))
}
}

\section{Arguments}{

\code{X.list} A list of data.frame

\code{methods} a vector of characters list the methods to be tested

\code{varname} The name of a Python variable to bring into R.
}

\section{Methods}{

\code{$new()} Initialize

\code{$roc.plot} plots
}

\examples{
\dontrun{
pypath = Sys.which('python')
py = PythonEnv$new(path = pypath, port = 6011, host = "127.0.0.1")
py$start()
py$running
py$set(a = 5)
py$get('a')
py$stop(force = TRUE)
}

}
