% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LITree-class.R
\name{graphModel}
\alias{graphModel}
\title{R6 Class for simulation of Graphs}
\description{
R6 Class for simulation of Graphs
}
\section{Usage}{

\preformatted{experiment = graphModel$new(type="erdos",size=30, p.or.m =0.1,eta=0.2,extraeta=eta/5)}
}

\section{Arguments}{


\code{- type="erdos"} type of simulated graph ('erdos', 'starerdos', 'star', 'tree')

\code{- size=NULL}  number of nodes (variables)

\code{- p.or.m=NULL}  probability of an edge (if smaller than 1) or number of edges (if greater than 1)

\code{eta=NULL}       factor affected to ...

\code{extraeta=NULL} factor affected  to ...

\code{adjmat=NULL}  simulated adjacency matrix
}

\section{Methods}{


\code{$new(type="erdos",size=30, p.or.m =0.1,eta=0.2,extraeta=eta/5)}  creates a random graph

\code{$plot()} plot the graph using the igraph plot function
}

\examples{
\dontrun{
star.graph <- graphModel$new(type = "starerdos",size=30, p.or.m = 0.05)
plot(star.graph)
}
}
