% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-LITree.R
\docType{package}
\name{LITree}
\alias{LITree}
\alias{LITree-package}
\title{LITree: A package for infering Gaussian Graphical Models with unobserved variables using latent tree agreggation}
\description{
The LITree package provides four classes  graphModel, GGMmodel, GGMfit,  GGMexperiment and one function em.litree()
}
\author{
Christophe Ambroise \email{christophe.ambroise@univ-evry.fr}

Geneviève Robin \email{genevieve.robin@polytechnique.edu}

Stéphane Robin \email{stéphane.robin@agroparistech.edu}
}
